--- src/shm_ugw_analysis/data_io/load.py
@@ -1,61 +1,0 @@
-import pathlib
-
-import numpy as np
-
-from shm_ugw_analysis.data_io.paths import NPZ_DIR
-
-
-def _load_npz(filename: pathlib.Path) -> tuple[np.ndarray, np.ndarray, np.ndarray]:
-    """Open the data from a file."""
-    contents = np.load(filename)
-    t, x, desc = contents['x'], contents['y'], contents['desc']
-    return x, t, desc
-
-
-allowed_cycles = ('AI', '0', '1', '1000', '10000', '20000', '30000', '40000', '50000', '60000', '70000', 'Healthy')
-allowed_signal_types = ('excitation', 'received')
-allowed_emitters = (1, 2, 3, 4, 5, 6)
-allowed_receivers = allowed_emitters
-allowed_frequencies = (100, 120, 140, 160, 180)
-
-
-def load_data(
-        cycle: str, signal_type: str, emitter: int, receiver: int, frequency: int
-) -> tuple[np.ndarray, np.ndarray, np.ndarray, str, str, pathlib.Path]:
-    """Loads the data for a given measurement."""
-    if cycle not in allowed_cycles:
-        raise ValueError(f"Invalid cycle '{cycle}', must be one of {allowed_cycles}")
-    if signal_type not in allowed_signal_types:
-        raise ValueError(f"Invalid signal type '{signal_type}, must be either 'excitation' or 'received'")
-    if emitter not in allowed_emitters:
-        raise ValueError(f"Invalid emitter number {emitter}, must be from 1 to 6")
-    if receiver not in allowed_receivers:
-        raise ValueError(f"Invalid receiver number {receiver}, must be from 1 to 6")
-    if emitter in [1, 2, 3]:
-        if receiver in [1, 2, 3]:
-            raise ValueError(f"Invalid combination of emitter number {emitter} and receiver number {receiver}")
-    elif receiver in [4, 5, 6]:
-        raise ValueError(f"Invalid combination of emitter number {emitter} and receiver number {receiver}")
-    if frequency not in allowed_frequencies:
-        raise ValueError(f"Invalid frequency '{frequency}', must be either 100, 120, 140, 160, or 180 [kHz]")
-
-    try:
-        folder = f'L2_S2_cycle_{int(cycle)}'
-    except ValueError:
-        folder = f'L2_S2_{cycle}'
-
-    c1_f1 = 'C1' if signal_type == 'excitation' else 'F1'
-    freq = str(frequency // 20)
-    if cycle == 'Healthy':
-        filename = f'{c1_f1}_L2_S2_healthy_emetteur_{emitter}_recepteur_{receiver}_excitation_{freq}_00000.npz'
-    else:
-        filename = f'{c1_f1}_mono_1_saine_emetteur_{emitter}_recepteur_{receiver}_excitation_{freq}_00000.npz'
-
-    full_path = NPZ_DIR.joinpath(folder, filename)
-
-    x: np.ndarray
-    t: np.ndarray
-    desc: np.ndarray
-
-    x, t, desc = _load_npz(full_path)
-    return x, t, desc, folder, filename, full_path
